refactor(Filter): 카드-필터 연결테이블 엔티티 생성
