+++ src/main/java/com/team/piickle/card/filter/CardFilter.java
@@ -1,0 +1,28 @@
+package com.team.piickle.card.filter;
+
+import com.team.piickle.card.domain.Card;
+import com.team.piickle.common.domain.BaseEntity;
+import javax.persistence.AttributeOverride;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AttributeOverride(name = "id", column = @Column(name = "CARD_FILTER_ID"))
+@Entity
+public class CardFilter extends BaseEntity {
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "CARD_ID")
+    private Card card;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "FILTER_ID")
+    private Filter filter;
+}
